【Pten】Refactor C++ API code-gen (#39408)
* refactor C++ API code-gen

* fix windows problem of C++ API
